move chart into separate component
