+++ src/components/Mortage.vue
@@ -69,22 +69,22 @@
             />
         </v-col>
         <v-col cols="8">
-            <apexchart ref="realtimeChart" width=620 :options="chartOptions" :series="series"></apexchart>
+            <chart :series="series"/>
         </v-col>
     </v-row>
 </div>
 </template>
 <script>
-import apexchart from 'vue-apexcharts'
+import Vue from 'vue'
+import chart from '@/components/Mortage/Chart'
 export default {
     components: {
-        apexchart
+        chart
     },
     data() {
         return {
             price: 500000,
             down: 100000,
-            displayDown:100000,
             downPercentage: 20,
             displayDownPercentage: 20,
             loanLength: 30,
@@ -94,68 +94,7 @@ export default {
             tax: 329,
             hoa: 100,
             changeFlag: false,
-            series:  [1756.23, 66, 329, 100],
-            
-            chartOptions: {
-                labels: ['Principal & interest', "Homeowner's insurance", 'Property tax', 'HOA fees'],
-                chart: {
-                    type: 'donut',
-                    animations: {
-                        speed: 480
-                    }
-                },
-                tooltip: {
-                    enabled: false
-                },
-                plotOptions: {
-                    pie: {
-                        donut: {
-                            labels: {
-                                show: true,
-                                name: {
-                                    show: true,
-                                    offsetY: -10,
-                                    formatter: function (val) {
-                                        return val+":"
-                                    }
-                                },
-                                value: {
-                                    show: true,
-                                    formatter: function (val) {
-                                        return "$"+val
-                                    }
-                                },
-                                total: {
-                                    show: true,
-                                    showAlways: false,
-                                    label: 'Monthly Payment',
-                                    fontSize: '20px',
-                                    fontFamily: 'Helvetica, Arial, sans-serif',
-                                    fontWeight: 550,
-                                    color: '#373d3f',
-                                    formatter: function (w) {
-                                        return "$"+w.globals.seriesTotals.reduce((a, b) => {
-                                        return Math.round((a + b) * 100.0) / 100.0;
-                                        }, 0)
-                                    }
-                                }
-                            }
-                        }
-                    }
-                },
-                responsive: [{
-                    options: {
-                        legend: {
-                        position: 'bottom'
-                        }
-                    }
-                }]
-            }
-        }
-    },
-    methods:{
-        updateSeriesLine() {
-            this.$refs.realtimeChart.updateSeries(this.series, false, true);
+            series:  [1756.23, 66, 329, 100]
         }
     },
     watch: {
@@ -190,29 +129,25 @@ export default {
             }
         },
         result: function(val) {
-            this.series[0] = val;
-            this.updateSeriesLine();
+            Vue.set(this.series, 0, val);
         },
         insurance: function(val) {
             if(val == "") {
                 val = 0;
             }
-            this.series[1] = val;
-            this.updateSeriesLine();
+            Vue.set(this.series, 1, val);
         },
         tax: function(val) {
             if(val == "") {
                 val = 0;
             }
-            this.series[2] = val;
-            this.updateSeriesLine();
+            Vue.set(this.series, 2, val);
         },
         hoa: function(val) {
             if(val == "") {
                 val = 0;
             }
-            this.series[3] = val;
-            this.updateSeriesLine();
+            Vue.set(this.series, 3, val);
         },
         price: function(val) {
             if(val == ""){
